Move Consumable and LiveData utils to shared module
