+++ audiorecorder/build.gradle
@@ -43,11 +43,12 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.activity:activity-ktx:1.2.0'
     implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.3.0'
+    implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.3.0"
     implementation project(':strings')
     implementation "com.google.dagger:dagger:${rootProject.daggerVersion}"
     implementation "com.jakewharton.timber:timber:${rootProject.timberVersion}"
     implementation project(path: ':async')
+    implementation project(path: ':shared')
     testImplementation project(path: ':testshared')
     kapt "com.google.dagger:dagger-compiler:${rootProject.daggerVersion}"
