+++ audiorecorder/src/main/java/org/odk/collect/audiorecorder/recording/AudioRecorder.kt
@@ -1,8 +1,8 @@
 package org.odk.collect.audiorecorder.recording
 
 import androidx.lifecycle.LiveData
-import org.odk.collect.audiorecorder.Consumable
 import org.odk.collect.audiorecorder.recorder.Output
+import org.odk.collect.shared.data.Consumable
 import java.io.File
 import java.io.Serializable
