+++ audiorecorder/src/main/java/org/odk/collect/audiorecorder/recording/internal/RecordingRepository.kt
@@ -2,11 +2,10 @@ package org.odk.collect.audiorecorder.recording.internal
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import org.odk.collect.audiorecorder.Consumable
 import org.odk.collect.audiorecorder.recording.RecordingSession
+import org.odk.collect.shared.data.Consumable
 import java.io.File
 import java.io.Serializable
-import java.lang.Exception
 
 internal class RecordingRepository {
