+++ audiorecorder/src/main/java/org/odk/collect/audiorecorder/testsupport/StubAudioRecorder.kt
@@ -2,10 +2,10 @@ package org.odk.collect.audiorecorder.testsupport
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import org.odk.collect.audiorecorder.Consumable
 import org.odk.collect.audiorecorder.recorder.Output
 import org.odk.collect.audiorecorder.recording.AudioRecorder
 import org.odk.collect.audiorecorder.recording.RecordingSession
+import org.odk.collect.shared.data.Consumable
 import java.io.File
 import java.io.Serializable
 
@@ -36,7 +36,7 @@ class StubAudioRecorder(private val stubRecordingPath: String) : AudioRecorder()
 
     private var isRecording = false
     private var failOnStart = false
-    private val currentSession = MutableLiveData<RecordingSession>(null)
+    private val currentSession = MutableLiveData<RecordingSession?>(null)
     private val failedToStart = MutableLiveData<Consumable<Exception?>>(Consumable(null))
 
     override fun isRecording(): Boolean {
