+++ collect_app/src/main/java/org/odk/collect/android/audio/AudioRecordingErrorDialogFragment.java
@@ -13,9 +13,9 @@ import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
 import org.odk.collect.android.R;
 import org.odk.collect.android.injection.DaggerUtils;
-import org.odk.collect.audiorecorder.Consumable;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
 import org.odk.collect.audiorecorder.recording.MicInUseException;
+import org.odk.collect.shared.data.Consumable;
 
 import javax.inject.Inject;
