+++ collect_app/src/main/java/org/odk/collect/android/formentry/FormEntryViewModel.java
@@ -18,8 +18,8 @@ import org.odk.collect.android.analytics.AnalyticsEvents;
 import org.odk.collect.android.exception.JavaRosaException;
 import org.odk.collect.android.formentry.audit.AuditEvent;
 import org.odk.collect.android.javarosawrapper.FormController;
-import org.odk.collect.audiorecorder.MutableNonNullLiveData;
-import org.odk.collect.audiorecorder.NonNullLiveData;
+import org.odk.collect.shared.livedata.MutableNonNullLiveData;
+import org.odk.collect.shared.livedata.NonNullLiveData;
 import org.odk.collect.utilities.Clock;
 
 import java.util.Objects;
