+++ collect_app/src/test/java/org/odk/collect/android/audio/AudioRecordingControllerFragmentTest.java
@@ -20,10 +20,10 @@ import org.odk.collect.android.injection.config.AppDependencyModule;
 import org.odk.collect.android.permissions.PermissionsChecker;
 import org.odk.collect.android.preferences.PreferencesProvider;
 import org.odk.collect.android.support.RobolectricHelpers;
-import org.odk.collect.audiorecorder.MutableNonNullLiveData;
 import org.odk.collect.audiorecorder.recorder.Output;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
 import org.odk.collect.audiorecorder.testsupport.StubAudioRecorder;
+import org.odk.collect.shared.livedata.MutableNonNullLiveData;
 import org.odk.collect.utilities.Clock;
 import org.robolectric.annotation.Config;
