+++ shared/src/main/java/org/odk/collect/shared/livedata/LiveDataUtils.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.utilities;
+package org.odk.collect.shared.livedata;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MediatorLiveData;
